fix(file-storage): address PR #4184 review + green the CI (cargo-deny/dylint/cfs)
Follow-up to the review + CI feedback on PR #4184. All local verification green:
cargo fmt/clippy/build clean, `cargo deny check` ok, `make cfs-validate` 0 errors
(1 unrelated oagw warning), full suite (file-storage + sdk) 0 failures.

CI failures fixed:
- cargo-deny (Security): the failure was RUSTSEC-2026-0204 (crossbeam-epoch
  0.9.18, transitive via rayon/pingora), not the S3 deps — bumped to 0.9.20 via
  `cargo update -p crossbeam-epoch` (Cargo.lock only; deny.toml unchanged; the
  new rusty-s3/quick-xml/s3s-fs licenses were already covered).
- Dylint (DE1101 tests-in-separate-files): the ~810-line inline
  `#[cfg(test)] mod tests` in src/bin/sidecar.rs exceeded the limit — extracted
  to src/bin/sidecar_tests.rs via `#[path]`, with Cargo.toml autobins=false +
  an explicit [[bin]] so the sibling test file isn't treated as a binary.
- CFS artifact validation: (a) new operations.md wasn't a recognized artifact
  type -> added to the [[ignore]] list; (b) file-storage's nested crate layout
  (gears/file-storage/file-storage/src) was never scanned for @cpt markers —
  added the missing Nested Gears/Tests/SDK codebase entries (mirroring
  gears/system/$system); (c) added @cpt-begin/@cpt-end traceability markers for
  the implemented multipart flow/algo/dod instruction IDs (report-part,
  native/offset part write, complete assembly, combine-hashes, dod markers) and
  kept genuinely-deferred p2 steps unchecked; (d) referenced ADR-0005 from
  DESIGN.md's Key ADRs table (ref-missing-from-kind) and regenerated its TOC.

Review comments addressed:
- Migration m20260706_000003 (policies unique index, item 2.4) now dedups
  pre-existing duplicate rows before CREATE UNIQUE INDEX (Postgres DELETE..USING
  by newest updated_at/policy_id; SQLite correlated EXISTS) so it can't fail on
  a DB that already carries duplicates from the pre-fix upsert race; migration
  test injects dups pre-migration and asserts success + one survivor + enforced
  uniqueness.
- Sidecar: an undersized/mismatched multipart part now returns 400 (client
  error), not 413 (413 stays only for genuine over-limit); stale doc comment on
  the usage_test multipart helper corrected.

Docs: corrected the stale EXECUTION LOG in IMPLEMENTATION_PLAN_TEMP.txt (8
commits, item 1.7/S3 done, 1.12 symmetry done + reporter still unwired, 2.12
downgraded to PARTIAL/blocked, push status). The (report_part size-forge,
sha256 dedup, IdempotencyInsert, authorize helper) review items were already in
the pushed history.
