+++ gears/file-storage/docs/ADR/0005-cpt-cf-file-storage-adr-s3-client-selection.md
@@ -14,10 +14,10 @@ date: 2026-07-07
   - [Consequences](#consequences)
   - [Confirmation](#confirmation)
 - [Pros and Cons of the Options](#pros-and-cons-of-the-options)
-  - [rusty-s3 (+ quick-xml) — chosen](#rusty-s3--quick-xml--chosen)
-  - [object_store (Apache Arrow, `aws` feature) — fallback](#object_store-apache-arrow-aws-feature--fallback)
-  - [aws-sdk-s3 (+ aws-config)](#aws-sdk-s3--aws-config)
-  - [rust-s3](#rust-s3)
+  - [`rusty-s3` (+ `quick-xml`) — chosen](#rusty-s3--quick-xml--chosen)
+  - [`object_store` (Apache Arrow, `aws` feature) — fallback](#object_store-apache-arrow-aws-feature--fallback)
+  - [`aws-sdk-s3` (+ `aws-config`)](#aws-sdk-s3--aws-config)
+  - [`rust-s3`](#rust-s3)
   - [Raw HTTP (no S3 client)](#raw-http-no-s3-client)
 - [More Information](#more-information)
   - [Option Comparison](#option-comparison)
