+++ gears/file-storage/docs/IMPLEMENTATION_PLAN_TEMP.txt
@@ -3532,12 +3532,12 @@ only has to happen once against the final schema rather than twice.
 
 ---
 
-## EXECUTION LOG — remediation run (2026-07-06/07)
+## EXECUTION LOG — remediation run (2026-07-06/07, updated 2026-07-07)
 
 > Appended after executing this plan. Branch: `feat/file-storage-p2-remediation`
 > (cut from `main` after P2 merge #4170; P2 is fully in `main`). Each step was
 > implemented as an isolated agent with build/clippy/test verification between
-> steps. Four DCO-signed commits, one per tier. Working tree clean.
+> steps. Eight DCO-signed commits. Working tree clean.
 >
 > NOTE: the Cargo package is `cf-gears-file-storage` (+ `-sdk`), not
 > `file-storage` — every `-p file-storage` command in this plan must use
@@ -3548,8 +3548,14 @@ only has to happen once against the final schema rather than twice.
 - `cd36794f6` — Tier 1 (11 of 14): production-readiness (blocked items excluded)
 - `f8d53f095` — Tier 2 (13/13): correctness hardening
 - `b66ff2573` — Tier 3 (11/11): documentation reconciliation
-- Total: 77 files, +11025 / -669. Full gear suite green (119 lib + integration +
-  sdk), `cargo fmt --check` + `cargo clippy --all-targets -D warnings` clean,
+- `1af3e1cb6` — PR #4184 review fixes (CodeRabbit)
+- `bd0ea1fd9` — 1.12 usage-accounting symmetry (credit stored bytes on
+  finalize/multipart-complete, debit on delete/sweep; zero-sum invariant test)
+- `587fc8079` — quota implementation-status docs + this execution log
+- `3153fd13c` — S3 storage backend (rusty-s3), item 1.7
+- Total: 77 files, +11025 / -669 through `587fc8079` (S3 commit adds further
+  files/lines on top). Full gear suite green (119 lib + integration + sdk),
+  `cargo fmt --check` + `cargo clippy --all-targets -D warnings` clean,
   sidecar builds, OpenAPI (`docs/api/api.json`) regenerated via `make openapi`.
 
 ### DONE — Tier 0 (all)
@@ -3559,28 +3565,51 @@ Migrations added: `m20260706_000001_idempotency_subject_id` (0.10).
 ### DONE — Tier 1 (cleanly-implementable subset)
 1.1, 1.2(a), 1.2(b), 1.3, 1.4(sweep half), 1.5, 1.6, 1.8, 1.9, 1.10(option b), 1.11.
 
-### DONE — Tier 2 (all)
+### DONE — 1.7 (S3 backend, landed after Tier 1-3 in commit `3153fd13c`)
+S3 client chosen: **rusty-s3** (+ `quick-xml`), with `object_store` kept as the
+documented fallback (ADR-0005). Fully implemented: `S3Backend` covering all
+`StorageBackend` ops including multipart and `put_stream`;
+`S3BackendConfig`/`s3_backends` config plus `default_backend_id`; `gear.rs`
+registry wiring; sidecar `claims.backend_id` dispatch to the S3 backend;
+s3s-fs-backed unit tests plus an optional live-s3s-fs e2e suite. Caveats that
+REMAIN open: the external-SDK security review of rusty-s3/quick-xml (ADR-0005
+stays `proposed`, not `accepted`), and Tier-4 real-object-store load/perf
+validation (unaffected by this landing).
+
+### DONE — Tier 2 (all, with 2.12 partial — see below)
 2.1 … 2.13. Migrations added: `m20260706_000002_idempotency_request_hash` (2.1),
 `m20260706_000003_policies_unique_scope` (2.4). New guardrail:
 `tests/error_mapping_test.rs` (exhaustive DomainError→status; adding a variant
 without a row is a compile error). `2.4` used the transaction + partial-unique-index
-backstop (not raw ON CONFLICT) to avoid bypassing SecureORM scope validation.
+backstop (not raw ON CONFLICT) to avoid bypassing SecureORM scope validation, and
+now also includes dedup/backfill of pre-existing duplicates: the migration's `UP`
+SQL deletes all-but-one row from every `(tenant_id, scope, scope_owner_id)` group
+before creating the unique indexes, so it doesn't fail against existing data.
+Note: **2.12 is PARTIAL, not fully done** — see the blocked-items list below.
 
-### DONE — Tier 3 (all)
+### DONE — Tier 3 (all shipped code, with 3.3/3.4 partial — see below)
 3.1 … 3.11. New `docs/operations.md`; `docs/migration.sql` synced; api.md /
 DESIGN.md / ADR-0003 / ADR-0004 / multipart-coordinator.md / README.md /
 DECOMPOSITION.md reconciled to the shipped code. New `DomainError::RetentionRuleNotFound`
-(3.10). InactivityRetention re-documented as "not modified" (3.11).
+(3.10). InactivityRetention re-documented as "not modified" (3.11). **3.3 and 3.4
+are PARTIAL**: docs were reconciled to reality, but 3.3's richer multipart-`complete`
+code contract and 3.4's GET-multipart-introspect endpoint remain deferred/tracked,
+not implemented — see the deferred-sub-parts list below.
 
 ### NOT DONE — blocked, needs a team decision / external dependency
-- **1.7 — S3 backend**: external-SDK security review + S3 crate choice + heavy new
-  dependency. Not started. Unblocks 0.2's native multipart and 1.2(b)/1.7.4 streaming.
-  S3 client selected: **rusty-s3** (primary), **object_store** (fallback) — see ADR-0005;
-  implementation still blocked on the security-review gate + SDK wiring.
 - **1.4 — quota half**: no `QuotaClient` SDK crate exists in the workspace.
   `gear.rs`'s `None, // quota_client` left as-is.
-- **1.12 — usage accounting symmetry**: coupled with wiring `usage_reporter` off
-  `None` (usage-collector-sdk exists) — needs a wiring decision; do in one change.
+- **1.12 — usage accounting, wiring half only**: the delta-symmetry fix itself
+  is DONE (commit `bd0ea1fd9`: credit stored bytes on finalize/multipart-complete,
+  debit on delete/sweep, zero-sum invariant test). What remains blocked is only
+  wiring `usage_reporter` off `None` — the usage-collector-sdk is not instantiated
+  in `gear.rs` yet; needs a wiring decision.
+- **2.12 — transfer_ownership, PARTIAL**: only nil-UUID target rejection was
+  implemented (cross-tenant transfer is already structurally impossible). The
+  item's DoD — reject nonexistent/unauthorized target owners — is NOT fully met:
+  full principal-existence/tenant-membership validation is blocked on an
+  account-management SDK, and whether transfer needs a distinct privileged
+  authorization action is an open 🛑 team decision (ties to 0.7's admin-scope).
 
 ### NOT DONE — Tier 4 (4.1–4.11): deferred/P3 by design
 Deliverable is a tracking ticket per item, not code — not filed by this run.
@@ -3628,5 +3657,7 @@ Deliverable is a tracking ticket per item, not code — not filed by this run.
   that's a `toolkit-canonical-errors` change (cross-cutting), not local.
 
 ### Process left
-- Push Tier 1–3 to update PR #4184 (per instruction, only Tier 0 was pushed).
+- Everything through `587fc8079` (Tier 0–3, PR-review fixes, 1.12 symmetry,
+  quota docs + this log) is already pushed to the fork and reflected on
+  PR #4184. Only the S3 commit `3153fd13c` remains unpushed.
 - Delete this file once the work is split into tracked tickets/PRs.
